requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf_yasg
Pillow